requirements.txt
# Be careful before running the installation, as some of the requirements depend on your device, such as PyTorch dependencies.
# My configuration may not work on your PC. Please check the requirements before running the installation.

# Without flux1schnell
streamlit
langchain
langchain_ollama
selenium
beautifulsoup4
lxml
html5lib
python-dotenv

#With flux1schnell
aesara==2.9.4
aim==3.25.0
aiohttp==3.10.9
amdsmi==6.2.1
apex==0.9.10dev
AppKit==0.2.8
astunparse==1.6.3
attr==0.3.2
audiocraft==1.3.0
av==13.1.0
azureml==0.2.7
beartype==0.19.0
beautifulsoup4==4.12.3
bokeh==3.6.0
boto3==1.35.37
brotli==1.1.0
brotlicffi==1.1.0.0
cairocffi==1.7.1
Carbon==1.1.10
causal_conv1d==1.4.0
checks==0.2
classy_vision==0.7.0
clearml==1.16.4
codecarbon==2.7.1
comet_ml==3.47.0
concurrencytest==0.1.2
conda==4.3.16
ConfigParser==7.1.0
contextlib2==21.6.0
coremltools==8.0
cryptography==3.4.8
ctypes_snappy==1.03
Cython==3.0.11
dagshub==0.3.39
dall_e==0.1
decord==0.6.0
deeplearning==0.0.1
defcon==0.10.3
defusedxml==0.7.1
dill==0.3.9
disco==1.40.4
dl==0.1.0
docutils==0.21.2
dp==1.00
dvclive==3.48.0
essentia==2.1b6.dev1177
etcd==2.0.8
evaluate==0.4.3
expecttest==0.2.1
fairseq==0.12.2
fairseq2==0.2.1
faiss==1.5.3
fastai==2.7.17
fastapi==0.115.0
fastparquet==2024.5.0
fb==0.4.0
fbgemm_gpu==0.8.0
flash_attn==2.6.3
flashlight==0.1.1
flatdict==4.0.1
flytekit==1.13.8
foo==.1
freetype_py==2.5.1
fs==2.4.16
ftfy==6.2.3
fugashi==1.3.2
g2p_en==2.1.0
galore_torch==1.0
gdown==5.2.0
gguf==0.10.0
gluonnlp==0.10.0
glyphsLib==6.9.2
gmpy2==2.2.1
gradio==5.0.1
h2==4.1.0
h5py==3.12.1
haiku==0.01
hf_transfer==0.1.8
HTMLParser==0.0.2
hypothesis==6.114.0
imageio==2.35.1
imwatermark==0.0.2
ini2toml==0.15
InquirerPy==0.3.4
intel_extension_for_pytorch==2.4.0
ipadic==1.0.0
ipykernel==6.29.5
ipython==8.12.3
ipywidgets==8.1.5
isal==1.7.1
jax==0.4.34
jaxlib==0.4.34
jieba==0.42.1
jnius==1.1.0
k_diffusion==0.1.1.post1
kenlm==0.2.0
keras_nlp==0.15.1
kerchunk==0.2.6
keyring==23.5.0
laion_clap==1.1.6
langchain_core==0.3.10
langchain_ollama==0.2.0
librosa==0.10.2.post1
lmdb==1.5.1
lomo_optim==0.1.1
lxml==5.3.0
lz4==4.3.3
lzmaffi==0.3.0
mamba_ssm==2.2.2
mambapy==1.2.0
Mask2Former==0.1.5
matchpy==0.5.5
megatron==0.5.1
minijinja==2.2.0
mlflow==2.16.2
mock==5.1.0
monkeytype==23.3.0
moviepy==1.0.3
mtrand==0.1
MultiScaleDeformableAttention==1.0
multiset==3.2.0
munkres==1.1.4
mxnet==1.9.1
mypy==1.11.2
natten==0.17.1
nemo==6.0.3
neptune==1.12.0
nltk==3.9.1
note_seq==0.0.5
olefile==0.47
onnxscript==0.1.0.dev20241010
optimum==1.22.0
optree==0.13.0
optuna==4.0.0
pandas==2.2.3
paramiko==3.5.0
phonemizer==3.3.0
pickle5==0.0.12
pippy==0.1.0
pkg1==0.0.3
pretty_midi==0.2.10
py==1.11.0
py3nvml==0.2.7
pyarrow==17.0.0
pycocotools==2.0.8
pycosat==0.6.6
pyctcdecode==0.5.0
pydantic==2.9.2
pydot==3.0.2
pydy==0.7.1
pygit2==1.15.1
pyglet==2.0.18
PyGObject==3.42.1
pygraphviz==1.14
PyInstaller==6.10.0
pymc==5.17.0
pymc3==3.11.6
pynvml==11.5.3
pyodide==0.0.2
pyOpenSSL==24.2.1
PyQt5==5.15.11
PyQt6==6.7.1
pysat==3.2.1
PySide2==5.15.2.1
PySide6==6.7.3
pytesseract==0.3.13
pythainlp==5.0.4
python_Levenshtein==0.26.0
pytorch_lightning==2.4.0
pytorch_quantization==2.2.1
pytz==2024.1
Quartz==0.0.1.dev0
railroad==0.5.0
ray==2.37.0
redis==5.1.1
reportlab==4.2.5
requests_kerberos==0.15.0
Res==0.1.7
rhoknp==1.7.0
rjieba==0.1.11
s3prl==0.4.16
sacremoses==0.1.1
sage==0.0.0
scann==1.3.4
scikit_learn==1.5.2
scipy==1.14.1
selenium==4.25.0
setuptools_scm==8.1.0
shiboken2==5.15.2.1
shiboken6==6.7.3
sigopt==8.8.3
simplejson==3.19.3
sip==6.8.6
slack_sdk==3.33.1
spacy==3.8.2
Sphinx==8.0.2
sphinx_gallery==0.18.0
starlette==0.39.2
streamlit==1.39.0
sudachipy==0.6.8
symengine==0.13.0
t5x==0.0.0
tabulate==0.9.0
tensorboardX==2.6.2.2
tensorflow_hub==0.16.1
tensorflow_probability==0.24.0
tensorflow_text==2.17.0
tensorstore==0.1.66
tf2onnx==1.16.1
tf_keras==2.17.0
theano==1.0.5
thread==2.0.5
threadpoolctl==3.5.0
tiktoken==0.8.0
timm==1.0.9
toml==0.10.2
tomli_w==1.1.0
torch_npu==2.3.1
torcharrow==0.1.0
torchrec==0.8.0
torchsde==0.2.6
tornado==6.4.1
trajectory==0.1.1
trove_classifiers==2024.9.12
ufoLib2==0.16.0
uharfbuzz==0.41.0
ujson==5.10.0
unicodedata2==15.1.0
unidic==1.1.0
unidic_lite==1.0.8
urllib3_secure_extra==0.1.0
uvicorn==0.31.1
vissl==0.1.6
wandb==0.18.3
wget==3.2
wmi==1.5.1
xattr==1.1.0
xformers==0.0.28.post1
xmlrpclib==1.0.1
xmlrunner==1.7.7
xx==3.3.2
yarl==1.14.0
z3==0.2.0
zopfli==0.2.3
zstandard==0.23.0
